Downloads/PAM/WestPoint/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
gspread==5.10.0
google-api-python-client==2.97.0
pandas==1.5.3
numpy==1.24.3
psycopg2==2.9.6
python-dotenv==1.0.0
